+++ src/lib/constants.js
@@ -88,7 +88,7 @@ module.exports = {
     ACTIVITY_RECORDER: {
         DECLARATION_TRACER: {
             MAX_RECORDS: 60,
-            PATH: '/shared/tmp/telemetry/declarationHistory'
+            PATH: '/var/log/restnoded/telemetryDeclarationHistory'
         }
     },
     APP_NAME: 'Telemetry Streaming',